Move inventory utils out of main class
